requirements.txt
Django==1.10.5
django-filter==1.0.1
psycopg2==2.6.2
